refactor(request_service): extract NetworkGuard from Policy
Move DNS resolution, host denylist, and blocked IP ranges into
NetworkGuard so Policy stays a thin public façade for limits and
validate_*! while SSRF/network rules keep a single construction home.

Callers (RuntimePolicy, Faraday, PuppetCommander) continue to talk only
to Policy — upholds Strong S3 single-homing of policy rules.
